requirements.txt
feedparser
trafilatura
pyyaml
pydantic
